refactor: TokenManager sharedPreferences 직접 주입 방식으로 코드 변경 https://github.com/Team-PATCH/Harmony-Android/pull/37#discussion_r1865939349
preferences 모듈 제거
