+++ core/data/build.gradle.kts
@@ -11,11 +11,12 @@ dependencies {
 
     implementation(project(":core:common"))
     implementation(project(":core:domain"))
-    implementation(project(":core:preferences"))
     implementation(project(":core:network"))
 
     implementation(libs.androidx.paging.runtime)
     implementation(libs.androidx.paging.compose)
+
+    implementation(libs.androidx.security.crypto)
 
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
