--- core/data/src/main/java/com/teampatch/core/data/repository/TokenRepositoryImpl.kt
@@ -1,17 +1,0 @@
-package com.teampatch.core.data.repository
-
-import com.teampatch.core.domain.repository.TokenRepository
-import com.teampatch.core.preferences.TokenLocalDataSource
-import javax.inject.Inject
-
-class TokenRepositoryImpl @Inject constructor(
-    private val tokenLocalDataSource: TokenLocalDataSource
-) : TokenRepository {
-    override fun getAccessToken(): String {
-        return tokenLocalDataSource.getAccessToken()
-    }
-
-    override fun setAccessToken(token: String) {
-        tokenLocalDataSource.setAccessToken(token)
-    }
-}
