--- core/domain/src/main/java/com/teampatch/core/domain/repository/TokenRepository.kt
@@ -1,7 +1,0 @@
-package com.teampatch.core.domain.repository
-
-interface TokenRepository {
-
-    fun getAccessToken(): String
-    fun setAccessToken(token: String)
-}
