--- core/network/src/main/java/com/teampatch/core/network/utils/TokenManager.kt
@@ -1,6 +1,0 @@
-package com.teampatch.core.network.utils
-
-interface TokenManager {
-
-    fun getAccessToken(): String
-}
