--- core/preferences/build.gradle.kts
@@ -1,17 +1,0 @@
-plugins {
-    id("teampatch.android.library")
-    id("teampatch.android.hilt")
-}
-
-android {
-    namespace = "com.teampatch.core.preferences"
-}
-
-dependencies {
-
-    implementation(libs.androidx.security.crypto)
-
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.espresso.core)
-}
