--- core/preferences/src/main/java/com/teampatch/core/preferences/TokenLocalDataSource.kt
@@ -1,10 +1,0 @@
-package com.teampatch.core.preferences
-
-interface TokenLocalDataSource {
-
-    fun getAccessToken(): String
-    fun setAccessToken(token: String)
-
-    fun getRefreshToken(): String
-    fun setRefreshToken(token: String)
-}
