--- core/preferences/src/main/java/com/teampatch/core/preferences/Utils.kt
@@ -1,3 +1,0 @@
-package com.teampatch.core.preferences
-
-internal const val FILE_NAME = "data-store.encrypted"
