+++ settings.gradle.kts
@@ -40,4 +40,3 @@ include(":feature:memorycard-registration")
 include(":core:data")
 include(":core:common")
 include(":core:network")
-include(":core:preferences")
